User Profile, Video Owner, User's Video, Access Control

+++ src/controller/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video"; //Using DB
+import User from "../models/User";
 const requestIp = require("request-ip");
 
 export const home = async (req, res) => {
@@ -12,7 +13,8 @@ export const home = async (req, res) => {
 export const watch = async (req, res) => {
   //영상페이지 로드 watch.pug
   const id = req.params.id;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner"); //User 데이터 Relationship
+  console.log(video);
   if (!video) {
     return res.render("404", { pageTitle: `Video Not Found` });
   }
@@ -22,10 +24,16 @@ export const watch = async (req, res) => {
 export const getEdit = async (req, res) => {
   //수정페이지 로드 edit.pug
   const id = req.params.id;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
 
   if (!video) {
     return res.render("404", { pageTitle: `Video Not Found` });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
 
   return res.render("edit", { pageTitle: `Edit : ${video.title}`, video });
@@ -34,11 +42,17 @@ export const getEdit = async (req, res) => {
 export const postEdit = async (req, res) => {
   //수정페이지 폼 POST edit.pug
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const { title, description, hashtags } = req.body;
   const video = await Video.exists({ _id: id }); //true or false
   if (!video) {
     //error check
     return res.status(404).render("404", { pageTitle: `Video Not Found` });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     //Update form
@@ -55,18 +69,25 @@ export const getUpload = (req, res) => {
 
 export const postUpload = async (req, res) => {
   //비디오 업로드 POST
+  const {
+    user: { _id },
+  } = req.session;
   const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body; //form 내용 받아오기
   try {
     //예외처리
-    await Video.create({
+    const newVideo = await Video.create({
       //form내용 DB저장
       title,
       fileUrl,
       description,
+      owner: _id,
       //createdAt:date.now(), << set default date
       hashtags: Video.formatHashtags(hashtags),
     }); //form 저장
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
     //에러 발생시
@@ -79,6 +100,17 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: `Video Not Found` });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
 
   return res.redirect("/");
